requirements.txt
playwright==1.52.0
psycopg[binary]==3.2.7
python-dotenv==1.0.1
